requirements.txt
#numpy==1.13
scipy==0.19.1
pandas==0.20.2
scikit_learn==0.18.2
Jinja2==3.0.1
MarkupSafe==2.0.1
Flask==2.0.3
Werkzeug==2.0.3
Pillow==8.4.0







